Fixed formatter crash on switch statements.
WARNING: This fix may not work with the new Java 12 switch syntax.

BACKGROUND:

The 2019-06 Eclipse changed the class heirarchy of SwitchStatement to
inherit from Expression instead of directly from Statement; this caused
the legacy Luna formatter to expect the switch block to end in a
semicolon. By ensuring that Expressions are not SwitchStatements, the
issue is mitigated.